Gaven concrete Inputs a no-arg constructor

+++ src/sudoku/Puzzle4x4.java
@@ -28,7 +28,7 @@ public class Puzzle4x4 extends SudokuPuzzle {
 	}
 	
 	private boolean rowIsValid(int row) {
-		short bits = 0;
+		byte bits = 0;
 		
 		for (int col = 0 ; col < 4 ; col++)
 			bits |= 1 << (super.board[row][col].value - '0' - 1);
@@ -37,7 +37,7 @@ public class Puzzle4x4 extends SudokuPuzzle {
 	}
 	
 	private boolean columnIsValid(int col) {
-		short bits = 0;
+		byte bits = 0;
 		
 		for (int row = 0 ; row < 4 ; row++)
 			bits |= 1 << (super.board[row][col].value - '0' - 1);
@@ -48,7 +48,7 @@ public class Puzzle4x4 extends SudokuPuzzle {
 	private boolean boxIsValid(int row, int col) {
 		int rowEnd = row + 2;
 		int colEnd = col + 2;
-		short bits = 0;
+		byte bits = 0;
 		
 		for (int row2 = row ; row2 < rowEnd ; row2++) {
 			for (int col2 = col ; col2 < colEnd ; col2++)
